requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
pillow>=10.0.0
pillow-heif>=0.10.0
opencv-python>=4.8.0

# Data Science & ML
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Metric Learning
pytorch-metric-learning>=2.0.0

# Training Utilities
tensorboard>=2.13.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0

# Utilities
pyyaml>=6.0
streamlit>=1.24.0
plotly>=5.15.0
